requirements.txt
wheel
discord.py
tortoise-orm==0.19.0
git+https://github.com/Mihitoko/discord.py-Console.git
git+https://github.com/Mihitoko/LangPy.git
aiocache~=0.11.1
coloredlogs~=15.0.1
pytz~=2022.1
aiohttp~=3.8.1
beautifulsoup4~=4.11.1
paramiko~=2.10.3
pydantic~=1.9.0
pypika-tortoise==0.1.4
asyncpg
lxml
yoyo-migrations